release: v0.4.0a1 preview
- Update pyproject.toml: bump version to 0.4.0a1, remove unused deps
  (Levenshtein, rich, logomaker), add min version constraints
- Update vampire.def: sync version label, add conda mirror support
  (Tsinghua TUNA for CN users), remove R packages
- Fix polars 1.x compatibility: add return_dtype to map_elements calls
- Update docs: installation --pre note, CLI reference (integrate/evaluate/
  refine/identity), news entry for 0.4.0a1 preview

+++ src/vampire/anno/pp/_read.py
@@ -455,7 +455,7 @@ def read_anno(
         motif_df.select([
             pl.col("id").cast(pl.Utf8),
             pl.col("motif"),
-            pl.col("motif").map_elements(lambda x: len(x)).alias("motif_length"),
+            pl.col("motif").map_elements(lambda x: len(x), return_dtype=pl.Int64).alias("motif_length"),
             pl.col("copyNumber").alias("copy_number"),
             pl.col("label"),
         ])
